Move DataMethod to client.internal.net
